refactor: move session schema to core
