+++ packages/core/src/db/index.ts
@@ -9,6 +9,7 @@ import type {
 export { coreSchema } from "./schema/shared";
 export { userSchema, type User } from "./schema/user";
 export { accountSchema, type Account } from "./schema/account";
+export { sessionSchema, type Session } from "./schema/session";
 
 export type {
 	DBFieldAttribute,
